backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
email-validator
mysql-connector-python
bcrypt
pyjwt